ttmp/2025-04-14/requirements.txt
requests==2.31.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-semantic-conventions==0.45b0 